+++ Ferreteria/Model/FacturaSD.cs
@@ -124,7 +124,7 @@ namespace Tienda.Model
             }
 
             // se le pasan los Aticulos  con sus detalles
-            public void AgregaArticulo(string Articulo, double precio, int cant, double subtotal)
+            public void AgregaArticulo(string Articulo, double precio, double cant, double subtotal)
             {
                 var can = cant.ToString().Length;
                 var pre = precio.ToString().Length;
